# packages/shared/openapi.yaml
openapi: 3.1.0
info:
  title: rescue-net.eu API
  version: 0.1.0
  description: >-
    Volunteer rescue alerting and mission coordination platform.
    This is the MVP skeleton contract. Endpoints are defined in the project
    manual section 14; only the implemented surface is documented here and the
    file grows as features land.
servers:
  - url: https://api.rescue-net.eu
  - url: http://localhost:8000
paths:
...
  /admin/retention/run:
    post:
      summary: Run GDPR retention sweeps on demand (platform admin; §16.2)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Counts per sweep
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: integer }
        "403": { description: Platform admin only }
  /admin/users/{user_id}/erase:
    post:
      summary: Erase a user's personal data (platform admin; §16.4)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Summary of removed personal data
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: integer }
        "404": { description: User not found }
# Realtime: connect to ws://<host>/ws/missions/{mission_id}?token=<jwt> to
# receive mission.* events (manual §14.2) fanned out over Redis pub/sub.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...